Capabilities drop support
Not that necessary as we're already confined in a user namespace,
but play it safe.

We could create a netns so that cap_net_raw has any effect (no caps
the fake root has apply to the netns it's running in) but that
requires a helper that has cap_net_admin in the upper netns to
connect the private netns to the outside world somehow.

For now, live with ping not working.