Use built-in tar to extract base images
This is possibly overkill (we're jumping into an empty root filesystem
to populate it from scratch) but who cares. Otheriwe we'd have to ignore
tar errors with /dev entries and possibly worry about symlinks and stuff.

tarfile comes with a nice default to ignore all errors, but we'll
probably have to up the errorlevel a bit (and provide our own error handler)