Choose a better target to restrict during download
The target root directory may not exist at this point, choose
either the Dockerfile base dir or the current directory

Target namespace's base dir would be also OK if guaranteed to be present,
as well as a completely empty directory or anything that definitely
exists.

Might as well create a temporary directory and remove it afterwards. TODO?